+++ R/RcppExports.R
@@ -202,14 +202,10 @@ xdev_add_references <- function(data, table, name = "name", vendor = "vendor", v
 #' @param data, a \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
 #'
 #' @param table, a data.frame containing your report.
-#'
-#' @param type, a string containing the type of report. Options include:
-#' "metadata" and custom report tags. Default = "metadata".
-#'
-#' @param sequence_name, a string containing the name of the column in 'table'
-#' that contains the sequence names. This is used for custom reports, metadata
-#' does not require a sequence_name column. Default column name is 'sequence_names'.
-#'
+#' @param type, a string containing the type of report. Default = "report".
+#' @param sequence_name, a string. If your report relates to the sequence data,
+#'   `sequence_name` should contain the name of the column in 'table' that
+#'   contains the sequence names. Default = 'none'.
 #' @param verbose, a boolean whether or not you want progress messages.
 #' Default = TRUE.
 #'
@@ -230,7 +226,7 @@ xdev_add_references <- function(data, table, name = "name", vendor = "vendor", v
 #'
 #' @return an updated \href{https://mothur.org/strollur/reference/strollur.html}{strollur} object
 #' @export
-xdev_add_report <- function(data, table, type = "metadata", sequence_name = "sequence_name", verbose = TRUE) {
+xdev_add_report <- function(data, table, type = "report", sequence_name = "none", verbose = TRUE) {
     .Call(`_strollur_xdev_add_report`, data, table, type, sequence_name, verbose)
 }
 
@@ -1112,9 +1108,9 @@ xdev_remove_sequences <- function(data, sequence_names, trash_tags) {
 #' @param type, string containing the type of report you would like. Options
 #' include: "fasta", "sequence", "sequence_bin_assignment",
 #' "sequence_taxonomy", "bin_taxonomy", "bin_representative",
-#'  "sample_assignment", "metadata", "resource_reference", "sequence_scrap",
+#'  "sample_assignment", "resource_reference", "sequence_scrap",
 #' "bin_scrap". If you have added custom reports for alignment,
-#' contigs_assembly or chimeras, you can get those as well.
+#' contigs_assembly, chimeras or metadata, you can get those as well.
 #'  Default = "sequence".
 #'
 #' @param bin_type, string containing the bin type you would like a bin_taxonomy