+++ NEWS.md
@@ -1,3 +1,8 @@
+# strollur 0.1.3
+
+* Adds `read_mothur_oligos()` function for the import of oligo data from [mothur](https://mothur.org). 
+* Makes report format more flexible. You can now add reports that are not tied to sequence data. These can be general metadata, oligo information, paired fastq file tables and much more.
+
 # strollur 0.1.2
 
 * Initial CRAN release.